move logger to utils folder
