+++ packages/core/src/game/objectives/check-team-objective.ts
@@ -1,4 +1,4 @@
-import { logger } from '../../auth/utils/logger';
+import { logger } from '../../utils/logger';
 import { Student } from '../../student/student.interface';
 import {
   updateStudent,
