+++ packages/core/src/pages/teacher/login.tsx
@@ -4,7 +4,7 @@ import { ReactElement, useState } from 'react';
 import { useAuth } from '../../auth/context/auth-context';
 import { UserRoles } from '../../auth/users/user-roles';
 import { clearAuthStorage } from '../../auth/utils/clear-auth-storage';
-import { logger } from '../../auth/utils/logger';
+import { logger } from '../../utils/logger';
 import { LoginForm } from '../../components/LoginForm';
 import { ApiHelper } from '../../api/api-helper';
