+++ packages/core/src/pages/teacher/teacher.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { getClassStats } from '../../admin/utils/get-class-stats';
 import { useAuth } from '../../auth/context/auth-context';
-import { logger } from '../../auth/utils/logger';
+import { logger } from '../../utils/logger';
 import { StorageKeys } from '../../auth/utils/storage-keys.constants';
 import { LoadingSpinner } from '../../components/LoadingSpinner';
 import { Modal } from '../../components/Modal';
